--- Tests/SwiftAvroCoreTests/AvroJSONEncoderGapTest.swift
@@ -1,340 +1,0 @@
-import Foundation
-import Testing
-@testable import SwiftAvroCore
-
-// Helpers private to this file (re-uses the same patterns as AvroJSONEncoderTest.swift).
-
-private func jsonAvro(_ schemaJSON: String) throws -> Avro {
-    let avro = Avro()
-    let schema = try #require(avro.decodeSchema(schema: schemaJSON))
-    avro.setSchema(schema: schema)
-    avro.setAvroFormat(option: .AvroJson)
-    return avro
-}
-
-// MARK: - Unkeyed primitive encode overloads with type mismatches
-
-@Suite("AvroJSONEncoder – unkeyed primitive type mismatches")
-struct UnkeyedPrimitiveMismatchTests {
-
-    private func wrongSchemaArray(items: String) throws -> Avro {
-        try jsonAvro(#"{"type":"array","items":"\#(items)"}"#)
-    }
-
-    @Test("Bool array against int schema throws")
-    func boolMismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([true, false]) }
-    }
-
-    @Test("Double array against int schema throws")
-    func doubleMismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([1.5, 2.5]) }
-    }
-
-    @Test("Float array against int schema throws")
-    func floatMismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([Float(1.0)]) }
-    }
-
-    @Test("Int (long-mapped) array against bool schema throws")
-    func intLongMismatch() throws {
-        let avro = try wrongSchemaArray(items: "boolean")
-        #expect(throws: (any Error).self) { try avro.encode([Int(1)]) }
-    }
-
-    @Test("Int8 array against long schema throws")
-    func int8Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "long")
-        #expect(throws: (any Error).self) { try avro.encode([Int8(1)]) }
-    }
-
-    @Test("Int16 array against long schema throws")
-    func int16Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "long")
-        #expect(throws: (any Error).self) { try avro.encode([Int16(1)]) }
-    }
-
-    @Test("Int32 array against long schema throws")
-    func int32Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "long")
-        #expect(throws: (any Error).self) { try avro.encode([Int32(1)]) }
-    }
-
-    @Test("Int64 array against int schema throws")
-    func int64Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([Int64(1)]) }
-    }
-
-    @Test("UInt array against int schema throws")
-    func uIntMismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([UInt(1)]) }
-    }
-
-    @Test("UInt16 array against long schema throws")
-    func uInt16Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "long")
-        #expect(throws: (any Error).self) { try avro.encode([UInt16(1)]) }
-    }
-
-    @Test("UInt32 array against int schema throws")
-    func uInt32Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([UInt32(1)]) }
-    }
-
-    @Test("UInt64 array against int schema throws")
-    func uInt64Mismatch() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([UInt64(1)]) }
-    }
-
-    @Test("UInt8 array against int schema throws")
-    func uInt8WrongSchema() throws {
-        let avro = try wrongSchemaArray(items: "int")
-        #expect(throws: (any Error).self) { try avro.encode([UInt8(1)]) }
-    }
-
-    @Test("[UInt8] against int (top-level non-bytes) throws")
-    func bytesArrayMismatch() throws {
-        // bytes guard at L693: only bytes schema accepts [UInt8]
-        let avro = try jsonAvro(#"{"type":"int"}"#)
-        #expect(throws: (any Error).self) { try avro.encode([UInt8(1), UInt8(2)]) }
-    }
-}
-
-// MARK: - Unkeyed string variants (string / enum / union)
-
-@Suite("AvroJSONEncoder – unkeyed string variants")
-struct UnkeyedStringVariantTests {
-
-    @Test("string array against enum schema (valid symbols)")
-    func enumValid() throws {
-        let avro = try jsonAvro(#"""
-        {"type":"array","items":{"type":"enum","name":"E","symbols":["X","Y","Z"]}}
-        """#)
-        let json = String(decoding: try avro.encode(["X", "Z"]), as: UTF8.self)
-        #expect(json.contains("\"X\""))
-        #expect(json.contains("\"Z\""))
-    }
-
-    @Test("string array against enum schema (invalid symbol throws)")
-    func enumInvalid() throws {
-        let avro = try jsonAvro(#"""
-        {"type":"array","items":{"type":"enum","name":"E","symbols":["X","Y"]}}
-        """#)
-        #expect(throws: (any Error).self) { try avro.encode(["NOPE"]) }
-    }
-
-    @Test("string array against union without string throws")
-    func unionWithoutStringThrows() throws {
-        let avro = try jsonAvro(#"{"type":"array","items":["null","int"]}"#)
-        #expect(throws: (any Error).self) { try avro.encode(["hi"]) }
-    }
-
-    @Test("string array against non-string non-enum non-union throws")
-    func defaultMismatch() throws {
-        let avro = try jsonAvro(#"{"type":"array","items":"int"}"#)
-        #expect(throws: (any Error).self) { try avro.encode(["text"]) }
-    }
-}
-
-// MARK: - Unkeyed bytes / fixed paths
-
-@Suite("AvroJSONEncoder – unkeyed bytes/fixed")
-struct UnkeyedBytesFixedTests {
-
-    @Test("[UInt8] against fixed schema (via generic encode<T>)")
-    func uint8ArrayAgainstFixed() throws {
-        let avro = try jsonAvro(#"{"type":"fixed","name":"F","size":4}"#)
-        let data = try avro.encode([UInt8]([0xDE, 0xAD, 0xBE, 0xEF]))
-        #expect(data.count > 0)
-    }
-
-    @Test("[UInt32] against fixed-duration schema (via generic encode<T>)")
-    func uint32ArrayAgainstFixedDuration() throws {
-        let avro = try jsonAvro(#"{"type":"fixed","name":"D","size":12,"logicalType":"duration"}"#)
-        let data = try avro.encode([UInt32]([1, 2, 3]))
-        let json = String(decoding: data, as: UTF8.self)
-        #expect(json.contains("1"))
-        #expect(json.contains("2"))
-        #expect(json.contains("3"))
-    }
-}
-
-// MARK: - Keyed container: [UInt8] / fixed / nested / superEncoder
-
-@Suite("AvroJSONEncoder – keyed bytes & nested coverage")
-struct KeyedBytesNestedTests {
-
-    @Test("record with [UInt8] field encoded as bytes")
-    func recordBytesField() throws {
-        struct R: Encodable { let payload: [UInt8] }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[
-          {"name":"payload","type":"bytes"}
-        ]}
-        """#)
-        let json = String(decoding: try avro.encode(R(payload: [0x01, 0x02, 0x03])), as: UTF8.self)
-        #expect(json.contains("\"payload\""))
-    }
-
-    @Test("record with [UInt8] field against non-bytes throws")
-    func recordBytesFieldMismatch() throws {
-        struct R: Encodable { let payload: [UInt8] }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[
-          {"name":"payload","type":"string"}
-        ]}
-        """#)
-        #expect(throws: (any Error).self) { try avro.encode(R(payload: [1, 2])) }
-    }
-
-    @Test("record with nested record uses keyed nestedContainer path")
-    func recordWithNestedRecord() throws {
-        // Exercise nestedContainer(keyedBy:forKey:) execution path. The
-        // encoder doesn't actually propagate child writes back to the parent
-        // for this manual path, so we only assert that the call doesn't
-        // throw — the goal is to execute L532–539 of AvroJsonEncoder.swift.
-        struct Outer: Encodable {
-            func encode(to encoder: Encoder) throws {
-                var k = encoder.container(keyedBy: AnyKey.self)
-                var nested = k.nestedContainer(keyedBy: AnyKey.self,
-                                               forKey: AnyKey(stringValue: "inner")!)
-                try nested.encode(Int32(7), forKey: AnyKey(stringValue: "a")!)
-            }
-        }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"Outer","fields":[
-          {"name":"inner","type":{"type":"record","name":"Inner","fields":[
-              {"name":"a","type":"int"}]}}
-        ]}
-        """#)
-        let data = try avro.encode(Outer())
-        #expect(data.count > 0)
-    }
-
-    @Test("record with nested array uses keyed nestedUnkeyedContainer path")
-    func recordWithNestedArray() throws {
-        // Exercises L541-548 (keyed nestedUnkeyedContainer creation). The
-        // returned child container has a schema mismatch when written
-        // manually, so we accept that the encode call may throw — the goal
-        // is to execute the construction path itself.
-        struct R: Encodable {
-            func encode(to encoder: Encoder) throws {
-                var k = encoder.container(keyedBy: AnyKey.self)
-                _ = k.nestedUnkeyedContainer(forKey: AnyKey(stringValue: "nums")!)
-            }
-        }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[
-          {"name":"nums","type":{"type":"array","items":"int"}}
-        ]}
-        """#)
-        _ = try? avro.encode(R())
-    }
-
-    @Test("encodeNil(forKey:) on null field succeeds")
-    func encodeNilOnNullField() throws {
-        struct R: Encodable {
-            func encode(to encoder: Encoder) throws {
-                var k = encoder.container(keyedBy: AnyKey.self)
-                try k.encodeNil(forKey: AnyKey(stringValue: "n")!)
-            }
-        }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[
-          {"name":"n","type":"null"}
-        ]}
-        """#)
-        let data = try avro.encode(R())
-        #expect(data.count > 0)
-    }
-
-    @Test("encodeNil(forKey:) on non-null non-union field throws")
-    func encodeNilThrowsOnPrimitive() throws {
-        struct R: Encodable {
-            func encode(to encoder: Encoder) throws {
-                var k = encoder.container(keyedBy: AnyKey.self)
-                try k.encodeNil(forKey: AnyKey(stringValue: "n")!)
-            }
-        }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[
-          {"name":"n","type":"int"}
-        ]}
-        """#)
-        #expect(throws: (any Error).self) { try avro.encode(R()) }
-    }
-
-    @Test("UInt64 too-large in record throws")
-    func uint64Overflow() throws {
-        struct R: Encodable { let v: UInt64 }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[{"name":"v","type":"long"}]}
-        """#)
-        #expect(throws: (any Error).self) { try avro.encode(R(v: UInt64.max)) }
-    }
-
-    @Test("UInt too-large in record throws")
-    func uintOverflow() throws {
-        struct R: Encodable { let v: UInt }
-        let avro = try jsonAvro(#"""
-        {"type":"record","name":"R","fields":[{"name":"v","type":"long"}]}
-        """#)
-        // On 64-bit platforms UInt.max > Int64.max
-        #expect(throws: (any Error).self) { try avro.encode(R(v: UInt.max)) }
-    }
-}
-
-// MARK: - Top-level array nestedContainer / nestedUnkeyedContainer
-
-@Suite("AvroJSONEncoder – unkeyed nested container creation")
-struct UnkeyedNestedContainerTests {
-
-    @Test("array of arrays via nestedUnkeyedContainer in unkeyed")
-    func arrayOfArrays() throws {
-        // Exercises L722-731 (unkeyed nestedUnkeyedContainer creation).
-        let avro = try jsonAvro(#"{"type":"array","items":{"type":"array","items":"int"}}"#)
-        struct AA: Encodable {
-            func encode(to encoder: Encoder) throws {
-                var outer = encoder.unkeyedContainer()
-                var inner = outer.nestedUnkeyedContainer()
-                try inner.encode(Int32(1))
-            }
-        }
-        let data = try avro.encode(AA())
-        #expect(data.count > 0)
-    }
-
-    @Test("array of records via nestedContainer in unkeyed")
-    func arrayOfRecordsViaNested() throws {
-        // Exercises L722-731 (unkeyed nestedContainer creation).
-        let avro = try jsonAvro(#"""
-        {"type":"array","items":{"type":"record","name":"R","fields":[
-          {"name":"x","type":"int"}]}}
-        """#)
-        struct AR: Encodable {
-            func encode(to encoder: Encoder) throws {
-                var outer = encoder.unkeyedContainer()
-                var nested = outer.nestedContainer(keyedBy: AnyKey.self)
-                try nested.encode(Int32(42), forKey: AnyKey(stringValue: "x")!)
-            }
-        }
-        let data = try avro.encode(AR())
-        #expect(data.count > 0)
-    }
-}
-
-// MARK: - Helper coding key
-
-private struct AnyKey: CodingKey {
-    var stringValue: String
-    var intValue: Int?
-    init?(stringValue: String) { self.stringValue = stringValue; self.intValue = nil }
-    init?(intValue: Int) { self.stringValue = "\(intValue)"; self.intValue = intValue }
-}
